pyproject.toml
[tool.poetry]
name = "research-assistant"
version = "0.1.0"
description = ""
authors = ["Samuel Halstead <worksam1114@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11.2"
pydantic = "^2.11.9"
pydantic-settings = "^2.10.1"
numpy = "^2.3.3"
fasttext-wheel = "^0.9.2"
chromadb = "^1.0.21"
openai = "^1.107.3"
fastapi = "^0.116.2"
uvicorn = "^0.35.0"

[tool.poetry.group.quality.dependencies]
pre-commit = "^4.3.0"
black = "~23.1.0"
isort = "~5.12.0"
mypy = "~1.0.1"
ruff = "~0.0.253"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
ignore-init-module-imports = true
line-length = 120
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    # "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]
